restructure row folder heirachy
