+++ packages/tables/src/lib/table/paginationControls/NumRowsControls.svelte
