+++ packages/tables/src/lib/table/paginationControls/PaginationControls.svelte
