+++ packages/tables/src/lib/table/rows/groupRows/GroupSummaryRow.svelte
@@ -1,10 +1,10 @@
-<script lang="ts">
-	import Scaffolding from './Scaffolding.svelte';
-	import ColGroupGap from '../cells/ColGroupGap.svelte';
+<script>
+	import ColGroupGap from '../../cells/ColGroupGap.svelte';
+	import Scaffolding from '../Scaffolding.svelte';
 
+	export let group;
 	export let table;
-
-	const DEFAULT_CELL_WIDTH = '100px';
+	export let tableSpec;
 </script>
 
 <Scaffolding {table}>
@@ -12,18 +12,21 @@
 		{#each table.columnSpec as col, i}
 			{#if !table.visibleFields || table.visibleFields.includes(col.short_label)}
 				<div
-					style:width={col.cell.width ?? DEFAULT_CELL_WIDTH}
+					style:width={col.cell.width ?? table.widths.defaultCell}
 					class="was-td"
 					style="flex-shrink: 0"
 				>
-					{#if col.cell && col.cell.axisRenderer}
+					{#if col.group && col.group.renderer}
 						<svelte:component
-							this={col.cell.axisRenderer}
+							this={col.group.renderer}
+							values={table.getValsForGroup(group, col.short_label)}
+							extent={table.extents[col.short_label]}
 							colorScale={table.scales[col.short_label]}
 							posScale={table.posScales[col.short_label]}
-							extent={table.extents[col.short_label]}
-							{...col.cell}
+							{...col.group}
 						/>
+
+						<!-- <td>{row[col.short_label]}</td> -->
 					{/if}
 				</div>
 			{/if}
