+++ packages/tables/src/lib/table/rows/headerRows/ColumnGroupHeadingRow.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { sum } from 'd3-array';
-	import Scaffolding from './Scaffolding.svelte';
-	import ColGroupGap from '../cells/ColGroupGap.svelte';
+	import ColGroupGap from '../../cells/ColGroupGap.svelte';
+	import Scaffolding from '../Scaffolding.svelte';
 
 	export let table;
